move About and Technologies components to the about folder
