+++ src/components/sections/about/technologies-item.tsx
